exposed assets, moved login to be exposed, allowed underscored var names
