+++ src/demo/containers/Demo.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Link } from 'react-router-dom'
 import logoSvg from '../../images/logo.svg'
 import CustomMarkdownView from '../../lib/containers/demo_components/CustomMarkdownView'
-import Login from '../../lib/containers/demo_components/Login'
+import Login from '../../lib/containers/Login'
 import UserFavorites from '../../lib/containers/demo_components/UserFavorites'
 import UserProfile from '../../lib/containers/demo_components/UserProfile'
 import UserProjects from '../../lib/containers/demo_components/UserProjects'
