+++ src/lib/index.tsx
@@ -5,7 +5,7 @@ import QueryWrapper from './containers/QueryWrapper'
 import StackedBarChart from './containers/StackedBarChart'
 import StaticQueryWrapper from './containers/StaticQueryWrapper'
 import SynapseTable from './containers/SynapseTable'
-import Login from './containers/demo_components/Login'
+import Login from './containers/Login'
 import SynapseTableCardView from './containers/SynapseTableCardView'
 import './style/Cards.css'
 import './style/Portal.css'
